+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 
 ### Changed
+- add maven coordinates to Library model
+- `version` is not part of the `mavenCoordinates` in the Library model
 
 ### Deprecated